rearrange(server): create routes /api, implement routes for /top-tracks
