+++ server/src/index.ts
@@ -3,6 +3,7 @@ import * as dotenv from "dotenv";
 import cors from "cors";
 import querystring from "querystring";
 import { PrismaClient } from '@prisma/client';
+import router from './routes';
 
 dotenv.config();
 
@@ -19,31 +20,9 @@ app.use(express.json());
 app.use(express.urlencoded({extended : true}));
 app.use(cors());
 
-let access_token: string | null = null;
-let expires_in: number | null = null;
-let refresh_token: string | null = null;
-
-interface Track {
-  available_markets: string[];
-  disc_number: number;
-  duration_ms: number;
-  explicit: boolean;
-  external_ids: {
-    isrc: string;
-  };
-  external_urls: {
-    spotify: string;
-  };
-  href: string;
-  id: string;
-  name: string;
-  popularity: number;
-  preview_url: string | null;
-  track_number: number;
-  type: string;
-  uri: string;
-  is_local: boolean;
-}
+export let access_token: string | null = null;
+export let expires_in: number | null = null;
+export let refresh_token: string | null = null;
 
 const refreshAccessToken = async () => {
   if (!refresh_token) {
@@ -195,229 +174,11 @@ app.get('/logout', (req, res) => {
   }
 });
 
-app.get('/top-tracks/short-term', async (req: Request, res: Response) => {
-  const { userId } = req.body;
-  // const userId = 5;
-  try {
-    // Fetch top tracks from Spotify API
-    const spotifyResponseJson = await fetch('https://api.spotify.com/v1/me/top/tracks?time_range=short_term&limit=20', {
-      headers: {
-        Authorization: `Bearer ${access_token}`,
-      },
-    });
-    const spotifyResponse = await spotifyResponseJson.json();
-    const spotifyTopTracks = spotifyResponse.items as Track[];
+app.use('/api', router);
 
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-    
-    const recentTopTracks = await prisma.shortTermTracks.findMany({
-      where: {
-        topTracksId: user?.topTracks?.id
-      },
-      orderBy: {
-        id: 'desc'
-      },
-      take: 1
-    });
-    const recentTopTracksData = recentTopTracks[0].trackData as string[];
-
-    const comparedTracks = spotifyTopTracks.map((spotifyTrack: Track, index) => {
-      const indexOfRecentTrack = recentTopTracksData?.indexOf(spotifyTrack.id);
-      if (indexOfRecentTrack === -1) {
-        return { change: 'new', ...spotifyTrack };
-      } else if (index === indexOfRecentTrack ) {
-        return { change: 'equal', ...spotifyTrack };
-      } else if (index < indexOfRecentTrack) {
-        return { change: 'up', ...spotifyTrack };
-      } else {
-        return { change: 'down', ...spotifyTrack };
-      }
-    });
-
-    res.status(200).json(comparedTracks);
-  } catch (error) {
-    console.error('Error fetching and comparing top tracks:', error);
-    res.status(500).json({ error: 'Failed to fetch and compare top tracks' });
-  }
-});
-
-app.post('/top-tracks/short-term', async (req: Request, res: Response) => {
-  try {
-    const { userId, tracks } = req.body;
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-
-    const createdTopTracks = await prisma.shortTermTracks.create({
-      data: {
-        topTracks: {
-          // connect: { id: userId },
-          connect: { id: user?.topTracks?.id },
-        },
-        trackData: tracks,
-        date: new Date().toLocaleDateString(),
-      },
-    });
-
-    res.status(201).json(createdTopTracks);
-  } catch (error) {
-    console.error('Error creating top tracks:', error);
-    res.status(500).json({ error: 'Failed to create top tracks' });
-  }
-});
-
-app.get('/top-tracks/medium-term', async (req: Request, res: Response) => {
-  const { userId } = req.body;
-  // const userId = 5;
-  try {
-    // Fetch top tracks from Spotify API
-    const spotifyResponseJson = await fetch('https://api.spotify.com/v1/me/top/tracks?time_range=medium_term&limit=30', {
-      headers: {
-        Authorization: `Bearer ${access_token}`,
-      },
-    });
-    const spotifyResponse = await spotifyResponseJson.json();
-    const spotifyTopTracks = spotifyResponse.items as Track[];
-
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-    
-    const recentTopTracks = await prisma.mediumTermTracks.findMany({
-      where: {
-        topTracksId: user?.topTracks?.id
-      },
-      orderBy: {
-        id: 'desc'
-      },
-      take: 1
-    });
-    const recentTopTracksData = recentTopTracks[0]?.trackData as string[] || [];
-
-    const comparedTracks = spotifyTopTracks.map((spotifyTrack: Track, index) => {
-      const indexOfRecentTrack = recentTopTracksData?.indexOf(spotifyTrack.id);
-      if (indexOfRecentTrack === -1 || recentTopTracksData?.length === 0) {
-        return { change: 'new', ...spotifyTrack };
-      } else if (index === indexOfRecentTrack ) {
-        return { change: 'equal', ...spotifyTrack };
-      } else if (index < indexOfRecentTrack) {
-        return { change: 'up', ...spotifyTrack };
-      } else {
-        return { change: 'down', ...spotifyTrack };
-      }
-    });
-
-    res.status(200).json(comparedTracks);
-  } catch (error) {
-    console.error('Error fetching and comparing top tracks:', error);
-    res.status(500).json({ error: 'Failed to fetch and compare top tracks' });
-  }
-});
-
-app.post('/top-tracks/medium-term', async (req: Request, res: Response) => {
-  try {
-    const { userId, tracks } = req.body;
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-
-    const createdTopTracks = await prisma.mediumTermTracks.create({
-      data: {
-        topTracks: {
-          // connect: { id: userId },
-          connect: { id: user?.topTracks?.id },
-        },
-        trackData: tracks,
-        date: new Date().toLocaleDateString(),
-      },
-    });
-
-    res.status(201).json(createdTopTracks);
-  } catch (error) {
-    console.error('Error creating top tracks:', error);
-    res.status(500).json({ error: 'Failed to create top tracks' });
-  }
-});
-
-app.get('/top-tracks/long-term', async (req: Request, res: Response) => {
-  const { userId } = req.body;
-  // const userId = 5;
-  try {
-    // Fetch top tracks from Spotify API
-    const spotifyResponseJson = await fetch('https://api.spotify.com/v1/me/top/tracks?time_range=long_term&limit=30', {
-      headers: {
-        Authorization: `Bearer ${access_token}`,
-      },
-    });
-    const spotifyResponse = await spotifyResponseJson.json();
-    const spotifyTopTracks = spotifyResponse.items as Track[];
-
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-    
-    const recentTopTracks = await prisma.longTermTracks.findMany({
-      where: {
-        topTracksId: user?.topTracks?.id
-      },
-      orderBy: {
-        id: 'desc'
-      },
-      take: 1
-    });
-    const recentTopTracksData = recentTopTracks[0]?.trackData as string[] || [];
-
-    const comparedTracks = spotifyTopTracks.map((spotifyTrack: Track, index) => {
-      const indexOfRecentTrack = recentTopTracksData?.indexOf(spotifyTrack.id);
-      if (indexOfRecentTrack === -1 || recentTopTracksData?.length === 0) {
-        return { change: 'new', ...spotifyTrack };
-      } else if (index === indexOfRecentTrack ) {
-        return { change: 'equal', ...spotifyTrack };
-      } else if (index < indexOfRecentTrack) {
-        return { change: 'up', ...spotifyTrack };
-      } else {
-        return { change: 'down', ...spotifyTrack };
-      }
-    });
-
-    res.status(200).json(comparedTracks);
-  } catch (error) {
-    console.error('Error fetching and comparing top tracks:', error);
-    res.status(500).json({ error: 'Failed to fetch and compare top tracks' });
-  }
-});
-
-app.post('/top-tracks/long-term', async (req: Request, res: Response) => {
-  try {
-    const { userId, tracks } = req.body;
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { topTracks: true },
-    });
-
-    const createdTopTracks = await prisma.longTermTracks.create({
-      data: {
-        topTracks: {
-          // connect: { id: userId },
-          connect: { id: user?.topTracks?.id },
-        },
-        trackData: tracks,
-        date: new Date().toLocaleDateString(),
-      },
-    });
-
-    res.status(201).json(createdTopTracks);
-  } catch (error) {
-    console.error('Error creating top tracks:', error);
-    res.status(500).json({ error: 'Failed to create top tracks' });
-  }
+app.use((error: Error, req: Request, res: Response, next: NextFunction) => {
+  console.error(error);
+  res.status(500).json({ error: 'Internal server error' });
 });
 
 app.listen(PORT, () => {
